+++ packages/framework/src/hooks/useEvaluate.ts
@@ -1,19 +1,49 @@
 import { useMemo } from "react";
-import { useAtomValue } from "jotai";
-import type { Expression } from "../shared";
-import { createBindingAtom } from "../state";
-import { useEnsembleStorage } from "./useEnsembleStorage";
+import { atom, useAtom } from "jotai";
+import { compact, merge, set } from "lodash-es";
+import { findExpressions } from "../shared";
+import { createBindingAtom } from "../evaluate";
 import { useCustomScope } from "./useCustomScope";
 
-export const useEvaluate = <T = unknown>(
-  expr: Expression<T>,
-): T | undefined | null => {
-  const storage = useEnsembleStorage();
+export const useEvaluate = <T extends Record<string, unknown>>(
+  values?: T,
+  options?: {
+    context?: unknown;
+    debugId?: string;
+  },
+): T => {
   const customScope = useCustomScope();
-  const valueAtom = useMemo(
-    () => createBindingAtom<T>(expr, { ensemble: { storage }, ...customScope }),
-    [customScope, expr, storage],
-  );
-  const value = useAtomValue(valueAtom);
-  return value;
+
+  const expressions = useMemo(() => {
+    const expressionMap: string[][] = [];
+    findExpressions(values, [], expressionMap);
+    return expressionMap;
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const bindingsAtom = useMemo(() => {
+    const bindingsEntries = compact(
+      expressions.map(([name, expr]) => {
+        const valueAtom = createBindingAtom(
+          expr,
+          merge({}, customScope, options?.context),
+          options?.debugId,
+        );
+        return { name, valueAtom };
+      }),
+    );
+    return atom((getAtom) => {
+      const valueEntries: [string, unknown][] = bindingsEntries.map(
+        ({ name, valueAtom }) => {
+          return [name, getAtom(valueAtom)];
+        },
+      );
+      const result = {};
+      valueEntries.forEach(([name, value]) => set(result, name, value));
+      return result;
+    });
+  }, [expressions, customScope, options?.context, options?.debugId]);
+
+  const [bindings] = useAtom(bindingsAtom);
+  return merge({}, values, bindings);
 };