+++ packages/runtime/src/widgets/Form/Checkbox.tsx
@@ -1,13 +1,15 @@
 import { Checkbox } from "antd";
-import { useState } from "react";
-import { useRegisterBindings } from "@ensembleui/react-framework";
+import { useMemo, useState } from "react";
+import { unwrapWidget, useRegisterBindings } from "@ensembleui/react-framework";
+import { isString } from "lodash-es";
 import type { EnsembleWidgetProps } from "../../shared/types";
+import { EnsembleRuntime } from "../../runtime";
 import { WidgetRegistry } from "../../registry";
 import type { FormInputProps } from "./types";
 import { EnsembleFormItem } from "./FormItem";
 
 export type CheckBoxProps = {
-  trailingText?: string;
+  trailingText?: string | Record<string, unknown>;
   leadingText?: string;
 } & EnsembleWidgetProps &
   FormInputProps<boolean>;
@@ -22,6 +24,16 @@ export const CheckboxWidget: React.FC<CheckBoxProps> = (props) => {
     },
   );
 
+  const trailingContent = useMemo(() => {
+    if (values?.trailingText) {
+      if (isString(values.trailingText)) {
+        return values.trailingText;
+      }
+
+      return EnsembleRuntime.render([unwrapWidget(values.trailingText)]);
+    }
+  }, [values?.trailingText]);
+
   return (
     <EnsembleFormItem valuePropName="checked" values={values}>
       <Checkbox
@@ -35,7 +47,7 @@ export const CheckboxWidget: React.FC<CheckBoxProps> = (props) => {
           ...values?.styles,
         }}
       >
-        {values?.trailingText}
+        {trailingContent}
       </Checkbox>
     </EnsembleFormItem>
   );